@@ -1,6 +1,7 @@
 {
   "name": "enso-studio-content",
   "version": "1.0.0",
+  "type": "module",
   "author": {
     "name": "Enso Team",
     "email": "contact@enso.org"
@@ -23,18 +24,18 @@
     "firebaseui": "^4.8.0",
     "html-loader": "^1.3.2",
     "mixpanel-browser": "2.40.1",
-    "@types/semver": "^7.3.5"
+    "@types/semver": "^7.3.9"
   },
   "devDependencies": {
     "compression-webpack-plugin": "^9.2.0",
-    "copy-webpack-plugin": "^10.2.4",
+    "copy-webpack-plugin": "^11.0.0",
     "node-polyfill-webpack-plugin": "^1.1.4",
-    "yaml-loader": "^0.6.0",
+    "yaml-loader": "^0.8.0",
     "source-map-loader": "^3.0.1",
-    "ts-loader": "^9.2.8",
-    "typescript": "^4.6.3",
-    "webpack": "^5.72.0",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.7.2",
+    "webpack": "^5.72.1",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "4.8.1"
+    "webpack-dev-server": "^4.9.0"
   }
 }
