@@ -61,7 +61,7 @@ ensogl::read_args! {
         authentication_enabled : bool,
         email                  : String,
         application_config_url : String,
-        /// When profiling the application (e.g. with the `./run.sh profile` command), this argument
+        /// When profiling the application (e.g. with the `./run profile` command), this argument
         /// chooses what is profiled.
         test_workflow            : String,
         skip_min_version_check   : bool,
